requirements.txt
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2